add resp3 protocol part 6(remove old protocol code) (#84)
